telepathy/farstreamchannel: Take parent state lock when changing state
according to gst docs: "Note that this is racy if the state lock of the
parent bin is not taken." when doing gst_element_set_locked_state(). That
is obviously the case here, leading to various races.

Fix that by taking element's parent lock before doing
gst_element_set_locked_state()

This seems to fix SIP calls not always having audio

Signed-off-by: Ivaylo Dimitrov <ivo.g.dimitrov.75@gmail.com>